refactor: move config classes into config folder
